backend/pyproject.toml
[project]
name = "backend"
version = "4.3.0"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "dj-database-url>=2.2.0",
    "django>=5.2.5",
    "django-celery-results>=2.5.1",
    "django-prometheus>=2.3.1",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.28.0",
    "gunicorn>=23.0.0",
    "meilisearch>=0.31.5",
    "pillow>=11.0.0",
    "sentry-sdk>=2.14.0",
    "whitenoise>=6.8.2",
    "django-utils-kit>=1.0.0",
    "django-meilisearch-indexer>=1.0.1",
    "django-anymail[sendinblue]>=13.0.1",
]
# ------------------------------
# UV
# ------------------------------
[tool.uv]
dev-dependencies = [
    "coverage>=7.6.4",
    "django-stubs>=5.1.1",
    "djangorestframework-stubs>=3.15.1",
    "factory-boy>=3.3.1",
    "ruff>=0.7.2",
    "tblib>=3.1.0",
    "ty>=0.0.1a3",
    "types-pytz>=2024.2.0.20241003",
    "types-requests>=2.32.0.20241016",
]
# ------------------------------
# TY
# ------------------------------
[tool.ty.rules]
unresolved-attribute = "ignore"
# ------------------------------
# RUFF
# ------------------------------
[tool.ruff.format]
quote-style = "double"
docstring-code-format = true
[tool.ruff.lint.isort]
known-first-party = [
    "core",
    "health",
    "django_react_starter",
    "user",
    "authentication",
]
force-sort-within-sections = true
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]